+++ tests/docs/test_docs_screenshots.py
@@ -77,7 +77,6 @@ def test_generate_documentation_screenshots(wizard, qtbot, docs_out, docs_tsv) -
         wizard, out / "clinical_scales_settings_dialog.png"
     )
     sh.save_electrode_canvas(s1.left_canvas, out / "electrode_diagram.png")
-    sh.save_electrode_canvas(s1.left_canvas, out / "electrode-canvas.png")
 
     qtbot.mouseClick(s1.next_button, Qt.MouseButton.LeftButton)
     sh.wait_for_render()